refactor: 외부에서 값을 주입받아 빈을 생성하도록 변경
